+++ src/main/java/jak/dut/Investment/model/calculation/CurrentInvestmentCalculation.java
@@ -1,9 +1,11 @@
 package jak.dut.Investment.model.calculation;
 
+import java.math.BigDecimal;
+
 public class CurrentInvestmentCalculation implements CalculaltionAlgorithmStrategy {
 
     @Override
-    public Double calculation(Calculation calculation) {
+    public BigDecimal calculation(Calculation calculation) {
         return null;
     }
 }
